+++ enterprise/internal/batches/resolvers/code_host_connection.go
@@ -12,32 +12,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/database"
 )
 
-type campaignsCodeHostConnectionResolver struct {
-	graphqlbackend.BatchChangesCodeHostConnectionResolver
-}
-
-var _ graphqlbackend.CampaignsCodeHostConnectionResolver = &campaignsCodeHostConnectionResolver{}
-
-func (c *campaignsCodeHostConnectionResolver) TotalCount(ctx context.Context) (int32, error) {
-	return c.BatchChangesCodeHostConnectionResolver.TotalCount(ctx)
-}
-
-func (c *campaignsCodeHostConnectionResolver) PageInfo(ctx context.Context) (*graphqlutil.PageInfo, error) {
-	return c.BatchChangesCodeHostConnectionResolver.PageInfo(ctx)
-}
-
-func (c *campaignsCodeHostConnectionResolver) Nodes(ctx context.Context) ([]graphqlbackend.CampaignsCodeHostResolver, error) {
-	batchNodes, err := c.BatchChangesCodeHostConnectionResolver.Nodes(ctx)
-	if err != nil {
-		return nil, err
-	}
-	nodes := make([]graphqlbackend.CampaignsCodeHostResolver, len(batchNodes))
-	for i, ch := range batchNodes {
-		nodes[i] = &campaignsCodeHostResolver{BatchChangesCodeHostResolver: ch}
-	}
-	return nodes, nil
-}
-
 type batchChangesCodeHostConnectionResolver struct {
 	userID                int32
 	onlyWithoutCredential bool
